add main file for testing and fixes
