+++ library/student.rb
@@ -1,0 +1,11 @@
+require_relative 'person'
+class Student < Person
+  def initialize(age, classroom, name = 'unknown', parent_permission: true)
+    super(age, name, parent_permission: parent_permission)
+    @classroom = classroom
+  end
+
+  def play_hooky
+    '¯\(ツ)/¯'
+  end
+end
