--- student.rb
@@ -1,11 +1,0 @@
-requiere_relative 'person'
-class Student < Person
-    def initialize(age, classroom, name ="Unknown", parent_permission : true)
-        super(age, name, parent_permission: parent_permission)
-        @classroom = classroom
-    end
-
-    def play_hooky
-        '¯\(ツ)/¯'
-    end
-end
